# spring-boot-hwsem6/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/reader": {
            "get": {
                "tags": [
                    "reader-controller"
                ],
                "operationId": "getAllReaders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reader"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "reader-controller"
                ],
                "operationId": "addNewReader",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reader"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reader"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/issue": {
            "post": {
                "tags": [
                    "issue-controller"
                ],
                "operationId": "issueBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IssueRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Issue"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/book": {
            "get": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "getAllBooks",
                "parameters": [
                    {
                        "name": "book",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "addBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reader/{id}": {
            "get": {
                "tags": [
                    "reader-controller"
                ],
                "operationId": "findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reader"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "reader-controller"
                ],
                "operationId": "deleteReader",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reader"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reader/{id}/issue": {
            "get": {
                "tags": [
                    "reader-controller"
                ],
                "operationId": "showAllIssue",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Issue"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/issue/{id}": {
            "get": {
                "tags": [
                    "issue-controller"
                ],
                "operationId": "getIssue",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Issue"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/book/{id}": {
            "get": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "book-controller"
                ],
                "operationId": "deleteBookById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Reader": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "IssueRequest": {
                "type": "object",
                "properties": {
                    "readerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "bookId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "Issue": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "idReader": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "idBook": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "dateOfTook": {
                        "type": "string"
                    },
                    "dateOfReturnBook": {
                        "type": "string"
                    }
                }
            },
            "Book": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            }
        }
    }
}